Add files via upload
Again, we ran into issues where the original data wasn't cleaned enough, and there were errors in the training set. We did not touch the test data, and we need to retrain the models on the cleaned data. The issue this time was that the BPM was completely unrealistic, and the mean was an absurdly high number. We cut out the outliers, because we found that the models were giving us MSE in the millions. Sorry again!